+++ src/parser/companions/DDBSummonsInterface.mjs
@@ -20,6 +20,7 @@ import { getClairvoyance } from "./types/Clairvoyance.mjs";
 import { getCreateUndead } from "./types/CreateUndead.mjs";
 import { getSpiritualWeapons } from "./types/SpiritualWeapon.mjs";
 import { getAnimateObjects2014 } from "./types/AnimateObjects.mjs";
+import { getPhantomSteed } from "./types/PhantomSteed.mjs";
 
 export default class DDBSummonsInterface {
 
@@ -66,5 +67,7 @@ export default class DDBSummonsInterface {
   static getSpiritualWeapons = getSpiritualWeapons;
 
   static getAnimateObjects2014 = getAnimateObjects2014;
+
+  static getPhantomSteed = getPhantomSteed;
 
 }