+++ src/parser/enrichers/spell/PhantomSteed.mjs
@@ -1,0 +1,45 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../data/DDBEnricherData.mjs";
+
+export default class PhantomSteed extends DDBEnricherData {
+
+  get type() {
+    return "summon";
+  }
+
+  get summonsFunction() {
+    return DDBImporter.lib.DDBSummonsInterface.getPhantomSteed;
+  }
+
+  get generateSummons() {
+    return !this.is2014;
+  }
+
+  get activity() {
+    return {
+      type: "summon",
+      noTemplate: true,
+      profileKeys: this.is2014
+        ? [
+          { count: 1, name: "PhantomSteed2014" },
+        ]
+        : [
+          { count: 1, name: "PhantomSteed2024" },
+        ],
+    };
+  }
+
+  get override() {
+    return {
+      data: {
+        flags: {
+          ddbimporter: {
+            disposition: {
+              match: true,
+            },
+          },
+        },
+      },
+    };
+  }
+}