+++ src/parser/enrichers/spell/_module.mjs
@@ -114,6 +114,7 @@ export { default as MirrorImage } from "./MirrorImage.mjs";
 export { default as Mislead } from "./Mislead.mjs";
 export { default as MistyStep } from "./MistyStep.mjs";
 export { default as PassWithoutTrace } from "./PassWithoutTrace.mjs";
+export { default as PhantomSteed } from "./PhantomSteed.mjs";
 export { default as Polymorph } from "./Polymorph.mjs";
 export { default as PowerWordFortify } from "./PowerWordFortify.mjs";
 export { default as PrimalSavagery } from "./PrimalSavagery.mjs";